+++ httpstatusfamily/Ranges.go
@@ -1,0 +1,5 @@
+package httpstatusfamily
+
+func Ranges() []string {
+	return ranges[:]
+}