+++ httpstatusfamily/all-is-checkers.go
@@ -1,0 +1,9 @@
+package httpstatusfamily
+
+func Is(rawString string, expected Variant) bool {
+	v, err := New(rawString)
+	if err != nil {
+		return false
+	}
+	return v == expected
+}